Optimize volatile-only fields in JVM
Atomic fields that use getValue/setValue are compiled down to the plain
volatile fields without FieldUpdater/VarHandle attached to them, so
that all @Volatile field in the source can be replaced with atomics
without loosing efficiency of the resulting bytecode.

In order to figure this out, method transformer is run in the Phase 2
on analysis in the special mode, so that data flow can be tracked
from fields to their method to see what methods are called on each
fields. Now transformation (Phase 3) is only run on files which this
Phase 2 had determined to need transformation (so we don't waste time
reading/reanalyzing class files that don't use atomics).

Also:
* File processing logic (phasing) is streamlined
* Array-handling logic is somewhat cleaned up