Add atomicfu dependency to all shared-native source sets
This fixes the cases when an HMPP project is built with atomicfu, and its shared-native source set is published with an atomicfu 'depends' entry in the manifest but without a proper transitive dependency on atomicfu. This leads to such an artifact not being usable as a dependency in other projects with shared-native source sets, as the Kotlin/Native compiler will reject it because of a missing transitive dependency.